+++ include/usolaris/vertex.hpp
@@ -5,8 +5,8 @@ namespace usolaris {
 
 // 頂点属性（圧縮フォーマット）
 struct Vertex {
-    _Float16 px, py, pz;  // 位置
-    int8_t u, v;           // UV     (/ 127.0f で [0,1] に展開)
+    float px, py, pz;     // 位置
+    int8_t u, v;          // UV     (/ 127.0f で [0,1] に展開)
     int8_t nx, ny, nz;    // 法線   (/ 127.0f で [-1,1] に展開)
     int8_t cr, cg, cb;    // 頂点色 (/ 127.0f で [0,1] に展開)
 };
